+++ src/stlink-gui/gui.c
@@ -262,7 +262,7 @@ static gpointer stlink_gui_populate_filemem_view(gpointer data) {
          */
 
         uint8_t* mem   = NULL;
-        size_t size    = 0;
+        uint32_t size    = 0;
         uint32_t begin = 0;
         int32_t res = stlink_parse_ihex(gui->filename, 0, &mem, &size, &begin);
