+++ src/win32/mmap.h
@@ -15,8 +15,8 @@
 #define MAP_ANONYMOUS (1 << 5)
 #define MAP_FAILED    ((void *)-1)
 
-void *mmap(void *addr, size_t len, int32_t prot, int32_t flags, int32_t fd, int64_t offset);
-int32_t munmap(void *addr, size_t len);
+void *mmap(void *addr, uint32_t len, int32_t prot, int32_t flags, int32_t fd, int64_t offset);
+int32_t munmap(void *addr, uint32_t len);
 
 #endif // STLINK_HAVE_SYS_MMAN_H
